+++ caffeine/src/test/java/com/github/benmanes/caffeine/cache/LoadingCacheTest.java
@@ -440,7 +440,7 @@ public final class LoadingCacheTest {
     LoadingCache<Object, Int> cache = context.build(key -> null);
 
     var keys = intern(new ArrayList<Key>());
-    for (int i = 0; i < Population.FULL.size(); i++) {
+    for (int i = 0; i < Math.toIntExact(Population.FULL.size()); i++) {
       keys.add(new Key());
     }
     Key key = requireNonNull(Iterables.getLast(keys));
